test(users): merge user api tests into users.routes.test.ts
Combine users.v2.api.test.ts and users.v2.avatar.api.test.ts, and add coverage for the migrated list, update and delete endpoints.
